Extract TST logic from Route to RouteNode
